Extracted smaller UI components from ThemeScreen for better readability
